feat: implement partial routing, HomePage, AboutPage, Navigation
