+++ src/components/App/index.css
@@ -1,64 +1,17 @@
 .app-container {
   display: flex;
   flex-direction: column;
-  text-align: center;
   min-height: 100vh;
+  background-color: var(--pokemon-white);
 }
 
-.header {
-  background: var(--dark-surface);
-  border-bottom: 1px solid var(--neon-blue);
+html {
+  scroll-behavior: smooth;
 }
 
-.app-container h1 {
-  font-family: 'Pokemon Solid', sans-serif;
-  color: var(--pokemon-yellow);
-  font-size: 3.5rem;
-  letter-spacing: 5px;
-  margin: 1.5rem 0;
-  text-shadow:
-    -4px -4px 0 var(--pokemon-black),
-    4px -4px 0 var(--pokemon-black),
-    -4px 4px 0 var(--pokemon-black),
-    4px 4px 0 var(--pokemon-black),
-    8px 8px 0 var(--pokemon-blue),
-    0 0 20px var(--pokemon-glow);
-  position: relative;
-  z-index: 1;
-}
-
-@media (max-width: 640px) {
-  .app-container h1 {
-    font-size: 2.5rem;
-    letter-spacing: 2px;
-  }
-}
-
-.main {
+.main-content {
   flex: 1;
-  padding: 2rem 1rem;
-  background-color: rgba(255, 255, 255, 0.9);
-  border-top: 3px solid var(--pokemon-black);
-  min-height: calc(100vh - 200px);
-  overflow-y: auto;
-}
-
-.error-button {
-  position: fixed;
-  right: 1.5rem;
-  bottom: 1.5rem;
-  padding: 0.75rem 1.5rem;
-  background: var(--pokemon-yellow);
-  color: var(--pokemon-black);
-  border: 3px solid var(--pokemon-black);
-  border-radius: 2rem;
-  font-weight: bold;
-  box-shadow: 2px 2px 0 var(--pokemon-black);
-  transition: all 0.2s ease;
-  z-index: 100;
-}
-
-.error-button:hover {
-  transform: translate(2px, 2px);
-  box-shadow: 0 0 0 var(--pokemon-black);
+  display: flex;
+  flex-direction: column;
+  padding-top: 60px;
 }
