+++ src/components/Navigation/index.css
@@ -1,0 +1,57 @@
+.navigation {
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  z-index: 1000;
+  display: flex;
+  justify-content: center;
+  gap: 0.5rem;
+  background: var(--dark-surface);
+  padding: 0.75rem 2rem;
+  border-bottom: 1px solid rgba(255, 255, 255, 0.1);
+  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.4);
+}
+
+.navigation a {
+  color: white;
+  text-decoration: none;
+  font-weight: 600;
+  font-size: 1rem;
+  padding: 0.5rem 1.25rem;
+  border-radius: 6px;
+  transition: all 0.2s ease;
+  text-transform: capitalize;
+  letter-spacing: 0.5px;
+  border: 1px solid transparent;
+  position: relative;
+  background: transparent;
+}
+
+.navigation a:hover {
+  background: rgba(255, 255, 255, 0.1);
+  color: var(--pokemon-yellow);
+  border-color: rgba(255, 255, 255, 0.2);
+  transform: translateY(-1px);
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.2);
+}
+
+.navigation a.active {
+  background: white;
+  color: var(--pokemon-red);
+  border-color: white;
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.2);
+  font-weight: 700;
+}
+
+@media (max-width: 640px) {
+  .navigation {
+    gap: 0.25rem;
+    padding: 0.5rem 1rem;
+  }
+
+  .navigation a {
+    font-size: 0.9rem;
+    padding: 0.4rem 1rem;
+  }
+}
