+++ src/components/SearchBar/index.css
@@ -1,11 +1,8 @@
-.search-bar-container {
-  margin: 2rem auto;
-  max-width: 42rem;
-}
-
 .search-bar {
   position: relative;
   width: 100%;
+  max-width: 600px;
+  margin: 0 auto;
 }
 
 .search-bar input {
@@ -14,7 +11,7 @@
   font-size: 1.1rem;
   padding: 1rem 9rem 1rem 1.5rem;
   color: var(--pokemon-black);
-  background: white;
+  background: var(--pokemon-white);
   border: 3px solid var(--pokemon-black);
   border-radius: 2rem;
   box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
