+++ src/pages/HomePage/index.css
@@ -1,0 +1,90 @@
+.home-page {
+  display: flex;
+  flex-direction: column;
+  text-align: center;
+  flex: 1;
+  min-height: 100%;
+  padding: 0 2rem;
+}
+
+.header {
+  box-sizing: border-box;
+  border-bottom: none;
+  padding: 2rem 0 1rem 0;
+  margin: 0 -32px;
+  height: 140px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.home-page h1 {
+  font-family: 'Pokemon Solid', sans-serif;
+  color: var(--pokemon-yellow);
+  font-size: 3.5rem;
+  letter-spacing: 5px;
+  margin: 0;
+  text-shadow:
+    -4px -4px 0 var(--pokemon-black),
+    4px -4px 0 var(--pokemon-black),
+    -4px 4px 0 var(--pokemon-black),
+    4px 4px 0 var(--pokemon-black),
+    8px 8px 0 var(--pokemon-blue),
+    0 0 20px var(--pokemon-glow);
+  position: relative;
+  z-index: 1;
+}
+
+@media (max-width: 640px) {
+  .home-page h1 {
+    font-size: 2.5rem;
+    letter-spacing: 2px;
+  }
+
+  .header {
+    height: 120px;
+  }
+
+  .home-page {
+    padding: 0 1rem;
+  }
+}
+
+.search-bar-container {
+  padding: 2rem;
+  margin: 0 -2rem;
+  position: relative;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+@media (max-width: 640px) {
+  .search-bar-container {
+    padding: 1.5rem 1rem;
+    margin: 0 -1rem;
+  }
+}
+
+.main {
+  flex: 1;
+  padding: 1rem 1rem 2rem 1rem;
+  background-color: var(--pokemon-beige-transparent);
+  border-top: none;
+  border-radius: 20px 20px 0 0;
+  margin-top: 0;
+  min-height: calc(100vh - 300px);
+  overflow-y: auto;
+  position: relative;
+  box-shadow: 0 -4px 15px rgba(0, 0, 0, 0.1);
+}
+
+.main::after {
+  content: '';
+  position: sticky;
+  bottom: 0;
+  height: 60px;
+  background: linear-gradient(to top, var(--pokemon-red), transparent);
+  margin: 2rem -1rem -2rem -1rem;
+  pointer-events: none;
+}
